requirements.txt
python-nmap
argcomplete
itertools
